requirements.txt
asgiref==3.6.0
Django==3.2.16
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.4.0
